Add additional `tracing` instrumentation.
This commit lays the ground-work for making mini-redis a better
example of an instrumented tokio application. While it does not
go so far to turn mini-redis into a fully-featured guide for
using `tracing`, it ensures that people who use mini-redis as a
basis for experimenting with different subscribers get a more
complete experience out-of-the-box.