rename feature flag in code, add CI job, fix cargo test
The code was still referencing the old feature flag name, when it
should have been using "otel". Also, added a new CI job that builds
using the "otel" flag. Also, add rand as a dependency because the
build was failing because `examples/load_test.rs` was using the `rand`
crate but it was not listed in the Cargo.toml dependencies